Написана логика для авторизации через ВК (не работает, развернуть local host)
