+++ src/shared/hooks/useVK/index.ts
@@ -5,12 +5,10 @@ import * as Crypto from 'expo-crypto';
 //https://id.vk.com/about/business/go/docs/ru/vkid/latest/vk-id/connection/start-integration/how-auth-works/auth-flow-web#Bez-SDK-s-obmenom-koda-na-bekende
 
 const vkAppId = String(process.env.EXPO_PUBLIC_VK_APP_ID);
-// const vkRedirectUri = AuthSession.makeRedirectUri({
-//     scheme: 'swans',
-//     path: 'continueWithVk',
-// });
-const vkRedirectUri =
-    'https://mmn6wqy-anonymous-8081.exp.direct/continueWithVk';
+let vkRedirectUri = AuthSession.makeRedirectUri({
+    scheme: 'swans',
+    path: 'continueWithVk',
+});
 
 type PKCE = {
     code_verifier: string;
@@ -84,64 +82,3 @@ const buildAuthorizeUrl = (params: {
     });
     return `${base}?${query.toString()}`;
 };
-
-function getQueryParams(url: string): Record<string, string> {
-    const params: Record<string, string> = {};
-    try {
-        const parsed = new URL(url);
-        parsed.searchParams.forEach((value, key) => {
-            params[key] = value;
-        });
-    } catch (e) {
-        console.warn('Failed to parse URL', e);
-    }
-    return params;
-}
-
-// import { Linking } from 'react-native';
-
-// const vkAppId = String(process.env.EXPO_PUBLIC_VK_APP_ID);
-// const vkRedirectUri =
-//     'https://mmn6wqy-anonymous-8081.exp.direct/continueWithVk';
-// const vkClientSecret = String(process.env.EXPO_PUBLIC_VK_CLIENT_SECRET);
-
-// export const handleVKAuth = async () => {
-//     const authUrl = `https://oauth.vk.com/authorize?client_id=${vkAppId}&display=mobile&redirect_uri=${encodeURIComponent(vkRedirectUri)}&scope=profile,email&response_type=code&v=5.131`;
-//     Linking.openURL(authUrl);
-// };
-
-// export const verifyVKAuth = async (queryParams: Record<string, string>) => {
-//     const { code } = queryParams;
-//     if (code) {
-//         try {
-//             const response = await fetch(
-//                 `https://oauth.vk.com/access_token?client_id=${vkAppId}&client_secret=${vkClientSecret}&redirect_uri=${encodeURIComponent(vkRedirectUri)}&code=${code}`,
-//             );
-//             const data = await response.json();
-//             console.log('Ответ от ВКонтакте:', data);
-
-//             if (data.error) {
-//                 console.error('Ошибка от ВКонтакте:', data.error);
-//                 return;
-//             }
-
-//             const { user_id, access_token, email } = data;
-//             if (user_id && access_token) {
-//                 console.log('Получены данные пользователя:', {
-//                     user_id,
-//                     access_token,
-//                     email,
-//                 });
-//             } else {
-//                 console.error('Не удалось получить данные пользователя');
-//             }
-//         } catch (error) {
-//             console.error(
-//                 'Ошибка при получении токена доступа ВКонтакте:',
-//                 error,
-//             );
-//         }
-//     } else {
-//         console.error('Код авторизации не найден');
-//     }
-// };
